apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
jinja2==3.1.3
openai==1.12.0
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.5.0
# Trending module dependencies
aiohttp==3.9.1
feedparser==6.0.10
apscheduler==3.10.4
pydantic-settings==2.1.0
slowapi==0.1.9
alembic==1.13.1
